Smal code clean-ups.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1534 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/lib/src/ColorBox.cpp
@@ -92,8 +92,8 @@ void ColorBox::setColor(const QColor& c)
 
 QColor ColorBox::color() const
 {
-	size_t i = this->currentIndex();
-	if (i >= 0 && i < sizeof(colors))
+        int i = this->currentIndex();
+        if (i >= 0 && i < (int)sizeof(colors))
 		return colors[this->currentIndex()];
 
 	return QColor(Qt::black); // default color is black.
